+++ src/comms.h
@@ -17,4 +17,4 @@ int comms_get_node_rank();
 void comms_barrier_procs();
 void comms_finalize_procs();
 void comms_init_procs(int *argc, char **argv[]);
-void comms_reduce_procs(double *sbuf, int count);
+void comms_reduce_procs(float *sbuf, int count);
